New adjustments

+++ R/coin_twitter.R
@@ -10,82 +10,42 @@ credentials <- function(file = "credentials.txt") {
     assign("token", TT, envir = .GlobalEnv)
   }
 
-load_tweets <- function(name, type = "user", format = "binary"){
+load_tweets <- function(name, type = c("user", "search", "stream"), format = c("binary","delimited")) {
+  if (format[1] == 'binary') extension <- ".dat" 
+  else  if (format[1] == "delimited") extension <- ".csv"
   
-  if (format == 'binary')
-  {
-    extension <- ".dat"
-  } else  if (format == "delimited")
-  {
-    extension <- ".csv"
-  }
+  if (type[1] == "user") output_file_name <- paste0("user_tweets_", name, extension)
+  else if (type[1] == "search") output_file_name <- paste0("search_tweets_", name, extension)
+  else if (type[1] == "stream") output_file_name <- paste0("stream_tweets_", name, extension)
   
-  if (type == "user")
-  {
-    output_file_name <- paste0("user_tweets_", name, extension)
-  }
-  else if (type == "search")
-  {
-    output_file_name <- paste0("search_tweets_", name, extension)
-  }
-  else if (type == "stream")
-  {
-    output_file_name <- paste0("stream_tweets_", name, extension)
-  }
-  
-  if (format == "binary")
-  {
-    load(output_file_name, .GlobalEnv )
-    
-  } else if (format == "delimited")
-  {
+  if (format[1] == "binary") load(output_file_name, .GlobalEnv)
+  else if (format[1] == "delimited") {
     tuits<-read_twitter_csv(file = output_file_name, unflatten = FALSE)
     return(tuits)
-    
   }
-  
-  
 }
 
-user_tweet <- function(user, maxtweets = 100, home = FALSE, parse = TRUE, check = TRUE, token = NULL, include_rts = FALSE, output_file_name = NULL, format = "binary"){
-  if (format == 'binary')
-  {
-    extension <- ".dat"
-  } else  if (format == "delimited")
-  {
-    extension <- ".csv"
-  }
+user_tweet <- function(user, maxtweets = 100, home = FALSE, parse = TRUE, check = TRUE, token = NULL, 
+                       include_rts = FALSE, output_file_name = NULL, format = c("binary", "delimited")) {
+  if (format[1] == 'binary') extension <- ".dat"
+  else  if (format[1] == "delimited") extension <- ".csv"
   
-  if(is.null(output_file_name)){
-    output_file_name <- paste0("user_tweets_", user, extension)
-  }
+  if(is.null(output_file_name)) output_file_name <- paste0("user_tweets_", user, extension)
   
   if(!(output_file_name %in% list.files())){
     datos_new <- get_timeline(user = user, n = maxtweets, home = home, parse = parse, check = check, token = token, include_rts = include_rts)
-    if (format == 'binary')
-    {
-      save( datos_new, file=output_file_name)
-    } else  if (format == "delimited")
-    {
-      write_as_csv(x = datos_new, file_name = output_file_name)
-    }
-    
-    
+    if (format[1] == 'binary') save( datos_new, file=output_file_name)
+    else  if (format[1] == "delimited") write_as_csv(x = datos_new, file_name = output_file_name)
     print(paste("Numero de tweets nuevos:", nrow(datos_new)))
     print("Nuevo fichero creado")
-  }else{
-    if (format == 'binary')
-    {
+  }
+  else {
+    if (format[1] == 'binary'){
       load(output_file_name, .GlobalEnv )
       datos_old <- datos_new
-      
-      
-    } else  if (format == "delimited")
-    {
-      datos_old<-read_twitter_csv(file = output_file_name, unflatten = TRUE)
-    }
-    
-    
+    } 
+    else  if (format[1] == "delimited") datos_old<-read_twitter_csv(file = output_file_name, unflatten = TRUE)
+
     ultimo_id <- datos_old[1, ]$status_id
     ultimo_id = toString(as.bigz(ultimo_id) + 1)
     datos_new <- get_timeline(user = user, n = maxtweets, since_id = ultimo_id,
@@ -93,42 +53,28 @@ user_tweet <- function(user, maxtweets = 100, home = FALSE, parse = TRUE, check 
     write_as_csv(x = datos_new, file_name = "./tmp.csv")
     datos_new<-read_twitter_csv(file = "./tmp.csv", unflatten = TRUE)
     tuits <- rbind(datos_new, datos_old)
-    if (format == "binary")
-    {
-      
-      save( tuits, file=output_file_name)
-      
-    } else if (format == "delimited")
-    {
-      write_as_csv(x = tuits, file_name = output_file_name)
-      
-    }
     
+    if (format[1] == "binary") save( tuits, file=output_file_name)
+    else if (format[1] == "delimited") write_as_csv(x = tuits, file_name = output_file_name)
     print(paste("Numero total de tweets:", nrow(tuits)))
     print(paste("Numero de tweets nuevos:", nrow(datos_new)))
   }
 }
 
-search_tweet <- function(search, maxtweets = 300, type = "recent", include_rts = TRUE, geocode = NULL, max_id = NULL, parse = TRUE, token = NULL, retryonratelimit = FALSE, verbose = TRUE, output_file_name = NULL){
-  if(is.null(output_file_name)){
-    output_file_name <- paste0("search_tweets_", search, ".csv")
-    
-  }
-  
-  if(!(output_file_name %in% list.files())){
+search_tweet <- function(search, maxtweets = 300, type = "recent", include_rts = TRUE, geocode = NULL, max_id = NULL, 
+                         parse = TRUE, token = NULL, retryonratelimit = FALSE, verbose = TRUE, output_file_name = NULL) {
+  if(is.null(output_file_name)) output_file_name <- paste0("search_tweets_", search, ".csv")
+  if(!(output_file_name %in% list.files())) {
     datos_new <- search_tweets(search, n = maxtweets, type = type, include_rts = include_rts, geocode = geocode, max_id = max_id, parse = parse, token = token, retryonratelimit = retryonratelimit, verbose = verbose)
     write_as_csv(x = datos_new, file_name = output_file_name)
     
     print(paste("Numero de tweets nuevos:", nrow(datos_new)))
     print("Nuevo fichero creado")
-  }else{
-    
+  }
+  else {
     datos_old<-read_twitter_csv(file = output_file_name, unflatten = TRUE)
-    
     ultimo_id <- datos_old[nrow(datos_old),]$status_id
-    
     ultimo_id = toString(as.bigz(ultimo_id) + 1)
-   
     datos_new <- search_tweets(search, n = maxtweets, type = type, include_rts = include_rts, geocode = geocode, max_id = ultimo_id, parse = parse, token = token, retryonratelimit = retryonratelimit, verbose = verbose)
     write_as_csv(x = datos_new, file_name = "./tmp.csv")
     datos_new<-read_twitter_csv(file = "./tmp.csv", unflatten = TRUE)
@@ -137,26 +83,17 @@ search_tweet <- function(search, maxtweets = 300, type = "recent", include_rts =
     print(paste("Numero total de tweets:", nrow(datos_concatenados)))
     print(paste("Numero de tweets nuevos:", nrow(datos_new)))
   }
-
 }
 
-stream_tweet <- function(search, timeout = 120, parse = TRUE, token = NULL, output_file_name = search, verbose = TRUE, dir = NULL){
+stream_tweet <- function(search, timeout = 120, parse = TRUE, token = NULL, output_file_name = search, verbose = TRUE, dir = NULL) {
     datos <- stream_tweets(search, timeout = timeout, parse = parse, token = token, file_name = output_file_name, verbose = verbose)
-    
     output_file_name_json <- paste0(output_file_name, ".json")
     datos_new <- parse_stream(output_file_name_json)
     output_file_name_csv <- paste0("stream_tweets_", output_file_name, ".csv")
     write_as_csv(x = datos_new, file_name = output_file_name_csv)
-    
     print(paste("Numero de tweets:", nrow(datos_new)))
     
-    
-    if (file.exists(output_file_name_json))
-    {
-      file.remove(output_file_name_json)
-    }
-  
-  
+    if (file.exists(output_file_name_json)) file.remove(output_file_name_json)
 }
 
 follow_tweet <- function(user, page = "-1", retryonratelimit = TRUE, parse = TRUE, verbose = TRUE, token = NULL){
@@ -168,43 +105,25 @@ follow_tweet <- function(user, page = "-1", retryonratelimit = TRUE, parse = TRU
   n <- scan()
   parte_entera <- floor(n/75000)
   parte_decimal <- (n/75000) - floor(n/75000)
-  if ((parte_entera == 0) | ((parte_entera == 1) & (parte_decimal == 0)))
-  {
-    partes <- 1
-  }
-  else
-  {
+  if ((parte_entera == 0) | ((parte_entera == 1) & (parte_decimal == 0))) partes <- 1
+  else {
     partes <- parte_entera
-    
     resto <- n - (partes * 75000)
-    if (parte_decimal > 0)
-    {
-      partes <- partes + 1
-    }
+    if (parte_decimal > 0) partes <- partes + 1
   }
-  if (partes == 1)
-  {
-    
-    seguidores_id <- get_followers(user, n, page = page, retryonratelimit = retryonratelimit, parse = parse, verbose = verbose, token = token)
-    
-  }
-  else
-  {
-    for (i in 1:partes)
-    {
-      if (i == 1)
-      {
+  if (partes == 1) seguidores_id <- get_followers(user, n, page = page, retryonratelimit = retryonratelimit, parse = parse, verbose = verbose, token = token)
+  else {
+    for (i in 1:partes) {
+      if (i == 1) {
         seguidores_id <- get_followers(user, n = 75000, page = page, retryonratelimit = retryonratelimit, parse = parse, verbose = verbose, token = token)
         next_page <- next_cursor(seguidores_id)
       }
-      else if (i > 1 & i < partes)
-      {
+      else if (i > 1 & i < partes) {
         seguidores_id2 <- get_followers(user, n = 75000, page = next_page, retryonratelimit = retryonratelimit, parse = parse, verbose = verbose, token = token)
         next_page <- next_cursor(seguidores_id2)
         seguidores_id <- rbind(seguidores_id, seguidores_id2)
       }
-      else if (i == partes)
-      {
+      else if (i == partes) {
         seguidores_id2 <- get_followers(user, n = resto, page = next_page, retryonratelimit = retryonratelimit, parse = parse, verbose = verbose, token = token)
         seguidores_id <- rbind(seguidores_id, seguidores_id2)
       }
@@ -231,37 +150,34 @@ prepare <- function(data=data, info=c("fields", "cotweets", "inform"), original=
   return(newdata)
 }
 
-prepareqda <- function(data=data)
-{
-  
-  newdata <- data.frame("Caso #" = "", "Caso" = "", "Variable" = "TWEET_TEXT", "PᲲafo" = "1", "No de resultados" = "0", "NO_CODES" = "")
-  newdata <- cbind (newdata, data %>% select(text, created_at, status_id, source, retweet_count, favorite_count, geo_coords))
+prepareqda <- function(data=data) {
+  newdata <- data.frame("Caso #" = "", "Caso" = "", "Variable" = "TWEET_TEXT", "Parrafo" = "1", "No de resultados" = "0", "NO_CODES" = "")
+  newdata <- cbind (newdata, data[,c("text", "created_at", "status_id", "source", "retweet_count", "favorite_count", "geo_coords")])
   names(newdata)[7:dim(newdata)[2]] <- c("Texto", "TWEET_DATE", "TWEET_ID", "SOURCE", "RETWEETS", "FAVORITES", "LONGITUDE")
   
-  newdata <- cbind (newdata, data %>% select(geo_coords, user_id, name, screen_name))
+  newdata <- cbind(newdata, data[, c("geo_coords", "user_id", "name", "screen_name")])
   
   names(newdata)[14:dim(newdata)[2]] <- c("LATITUDE", "USER_ID", "USER_NAME", "SCR_NAME")
   
-  newdata <- cbind (newdata, data %>% select(location, description, status_url, followers_count, friends_count, listed_count, account_created_at, favourites_count, 
-                                             country, statuses_count, profile_image_url, name, is_retweet))
+  fields <- c("location", "description", "status_url", "followers_count", "friends_count", "listed_count", "account_created_at", "favourites_count", 
+              "country", "statuses_count", "profile_image_url", "name", "is_retweet")
+  
+  newdata <- cbind(newdata, data[,fields])
+  
   names(newdata)[18:dim(newdata)[2]] <- c("LOCATION", "DESC", "URL", "FOLLOWERS", "FRIENDS", "LISTED", "USER_DATE", "USER_FAV", "TIMEZONE","TWEETS", "IMAGE", "NOMBRE", "T_RT")
   
   newdata$TIMEZONE <- "null"
   
-  
   newdata$T_RT <- ifelse(newdata$T_RT, "RT", "Tweet")
   
-  newdata$Caso.. <-row.names(newdata)
-  newdata$Caso <- paste("Case: ", row.names(newdata))
+  newdata$Caso.. <- row.names(newdata)
+  newdata$Caso   <- paste("Case: ", row.names(newdata))
  
   
-  for (i in 1:nrow(newdata))
-  {
+  for (i in 1:nrow(newdata)) {
     newdata[i,13] <- ifelse(newdata[i,13] == "NA NA", "NA", strsplit(newdata[i,13]," ")[[1]][1])
-    newdata[i,14] <-  ifelse(newdata[i,14] == "NA NA", "NA", strsplit(newdata[i,14]," ")[[1]][2])
+    newdata[i,14] <- ifelse(newdata[i,14] == "NA NA", "NA", strsplit(newdata[i,14]," ")[[1]][2])
   }
-  
-  
   return(newdata)
 }
